requirements.txt
# Core ML/NLP Dependencies
transformers>=4.35.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Retrieval Dependencies
rank-bm25>=0.2.2
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Evaluation Dependencies
evaluate>=0.4.0
scikit-learn>=1.3.0

# Observability
langfuse>=2.0.0

# LLM Integration
openai>=1.0.0
anthropic>=0.7.0

# Configuration Management
pyyaml>=6.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
types-PyYAML>=6.0.0
types-requests>=2.31.0

# HuggingFace
datasets>=2.14.0
huggingface-hub>=0.17.0
